Refactor: move encoded_sequence to count_kmers_specific dir
