+++ src/count_kmers_specific/encoded_sequence/encoded_sequence_proxy.h
